v0.2/requirements.txt
pandas
yfinance
lxml
beautifulsoup4
google-cloud-bigquery
flask
gunicorn
google-cloud-language